moved LogicContext to separate file
