+++ src/components/LogicContext.js
@@ -1,0 +1,3 @@
+import React from "react";
+const LogicContext = React.createContext(undefined);
+export default LogicContext
